Move request signing functions to new source file.
